# backend/src/main/webapp/openapi.yaml
openapi: 3.0.1
info:
  title: Musicshop API
  contact:
    name: Musicshop API Support
    url: http://musicshop.com/contact
    email: techsupport@musicshop.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 3.0.1
servers:
- url: http://10.0.40.166:8080/
  variables: {}
paths:
...
  /rest/playlist:
    get:
      summary: Retrieve the user playlist which contains all bought mediums
      operationId: getUserPlaylist
      responses:
        "200":
          description: Returns List<AlbumDTO>
        "401":
          description: Unauthorized
      security:
      - Authentication: []
...
components:
...
  securitySchemes:
    Authentication:
      type: http
      description: JWT token
      in: header
      scheme: bearer
      bearerFormat: JWT